+++ static/scripts/cirip/query-graph.ts
@@ -1,0 +1,12 @@
+export async function queryGraph(endpoint, query) {
+  const response = await fetch(endpoint, {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+      Accept: "application/json",
+    },
+    body: JSON.stringify({ query }),
+  });
+
+  return response.json();
+}
